+++ src/core/client.ts
@@ -1,3 +1,4 @@
+import type { Resource } from "./resource";
 import { createDebug } from "obug";
 import {
 	type ClientConfiguration,
@@ -47,7 +48,6 @@ import {
 	PARAMETER_DEFAULTS,
 	PARAMETERS,
 } from "./metric";
-import type { Resource } from "./resource";
 import { Sensor, Sensors } from "./sensor";
 import {
 	cleanKey,
@@ -380,9 +380,9 @@ export abstract class Client<
 		try {
 			const body = refreshToken
 				? JSON.stringify({
-						grant_type: "refresh_token",
-						refresh_token: refreshToken,
-					})
+					grant_type: "refresh_token",
+					refresh_token: refreshToken,
+				})
 				: undefined;
 
 			const res = await fetch(url, {
@@ -499,12 +499,12 @@ export abstract class Client<
 		let dt =
 			this.datetimeType === "string"
 				? new Datetime(dtValue as string, {
-						format: this.datetimeFormat,
-						timezone: this.timezone,
-					})
+					format: this.datetimeFormat,
+					timezone: this.timezone,
+				})
 				: new Datetime(toUnixSeconds(dtValue, this.datetimeType), {
-						locationTimezone: this.timezone,
-					});
+					locationTimezone: this.timezone,
+				});
 
 		if (!this.timeEnding) {
 			if (!averagingIntervalSeconds) {
@@ -629,13 +629,13 @@ export abstract class Client<
 
 		return this.normalizeDataStructure(
 			d as
-				| Partial<
-						Record<
-							"measurements" | "locations" | "meta" | "flags" | "sensors",
-							SourceRecord[]
-						>
-				  >
-				| SourceRecord[],
+			| Partial<
+				Record<
+					"measurements" | "locations" | "meta" | "flags" | "sensors",
+					SourceRecord[]
+				>
+			>
+			| SourceRecord[],
 		);
 	}
 
@@ -866,9 +866,9 @@ export abstract class Client<
 		const params: Array<
 			string | PathExpression | ConstantValue | ParseFunction
 		> = this.longFormat
-			? // for long format we will just pass the parameter name key and use that each time
+				? // for long format we will just pass the parameter name key and use that each time
 				[this.parameterName]
-			: this.measurements.parameterKeys();
+				: this.measurements.parameterKeys();
 
 		measurements.forEach((measurementRow: SourceRecord) => {
 			try {
@@ -1030,8 +1030,7 @@ export abstract class Client<
 		}
 
 		throw new Error(
-			`Invalid parser method: ${JSON.stringify(method)}${
-				key ? ` with key "${key}"` : ""
+			`Invalid parser method: ${JSON.stringify(method)}${key ? ` with key "${key}"` : ""
 			}`,
 		);
 	}
@@ -1085,8 +1084,7 @@ export abstract class Client<
 		}
 
 		throw new Error(
-			`Invalid reader method: ${JSON.stringify(method)}${
-				key ? ` with key "${key}"` : ""
+			`Invalid reader method: ${JSON.stringify(method)}${key ? ` with key "${key}"` : ""
 			}`,
 		);
 	}