+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [O.6.0] - unreleased
+
+### Added
+- LICENSE and NOTICE for Apache 2.0 license
+
+### Changed
+- `now` static method for `Datetime` to support `timezone` parameter.
+
 ## [0.5.0] - 2026-08-19
 
 ### Added